feat: refactored migrations and seeders

+++ database/migrations/2025_10_28_121804_create_grave_of_deceased_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('grave_of_deceased', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger("deceased_id");
-            $table->unsignedBigInteger("grave_id");
+            $table->integer("deceased_id");
+            $table->integer("grave_id");
             $table->timestamps();
         });
     }
